projects/actor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "navi-actor"
version = "0.1.0"
description = "Layer 3: The Brain — Cognitive Actor with Mamba2 SSM, PPO, and RND"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "navi-contracts>=0.1.0",
    "pyzmq>=26.0",
    "typer>=0.15",
    "torch>=2.2",
    "numpy>=1.26",
    "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
navi-actor = "navi_actor.cli:app"

[tool.uv.sources]
navi-contracts = { path = "../contracts", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/navi_actor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["--strict-markers", "--strict-config", "-ra", "--tb=short"]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disable_error_code = ["type-arg", "no-any-return", "unused-ignore"]

[[tool.mypy.overrides]]
module = [
    "zmq",
    "typer",
    "matplotlib",
    "matplotlib.*",
    "navi_contracts",
    "navi_contracts.*",
    "torch",
    "torch.*",
    "mamba_ssm",
    "mamba_ssm.*",
    "causal_conv1d",
    "causal_conv1d.*",
    "faiss",
    "faiss.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["navi_actor.cli"]
disable_error_code = ["misc"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
